+++ frontend/src/pages/api/account/login.js
@@ -22,9 +22,6 @@ export default async (req, res) => {
             });
 
             const data = await apiRes.json();
-            console.log(apiRes);
-            console.log("login: "+data +" "+ apiRes.status);
-            console.log("required method : "+ req.method);
 
             if (apiRes.status === 200) {
                 res.setHeader('Set-Cookie', [
