+++ frontend/src/pages/api/account/user.js
@@ -22,10 +22,6 @@ export default async (req, res) => {
             });
             const data = await apiRes.json();
 
-            console.log(apiRes);
-            console.log("load user: "+ data +" "+ apiRes.status);
-            console.log("required method : "+ req.method);
-
             if(apiRes.status === 200){
                 return res.status(200).json({
                     user: data
